+++ src/main/java/io/cdap/plugin/dynamo/sink/DynamoDBRecordWriter.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.cdap.plugin;
+package io.cdap.plugin.dynamo.sink;
 
 import com.amazonaws.services.dynamodbv2.document.BatchWriteItemOutcome;
 import com.amazonaws.services.dynamodbv2.document.DynamoDB;
@@ -23,8 +23,8 @@ import com.amazonaws.services.dynamodbv2.document.internal.InternalUtils;
 import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughputExceededException;
 import com.amazonaws.services.dynamodbv2.model.WriteRequest;
-import io.cdap.plugin.utils.DynamoDBConstants;
-import io.cdap.plugin.utils.DynamoDBUtil;
+import io.cdap.plugin.dynamo.sink.utils.DynamoDBConstants;
+import io.cdap.plugin.dynamo.sink.utils.DynamoDBUtil;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.mapreduce.RecordWriter;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
