+++ src/test/java/io/cdap/plugin/dynamo/sink/DynamoDBBatchSinkConfigTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.cdap.plugin;
+package io.cdap.plugin.dynamo.sink;
 
 import io.cdap.cdap.etl.mock.common.MockPipelineConfigurer;
 import org.junit.Assert;
